impl LineageContext with helper functions ot improve readability and logging
